sample/pyproject.toml
[project]
name = "sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.2"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
]
